apps/api/pyproject.toml
[project]
name = "sustainmetric-api"
version = "0.2.0"
description = "Sustainmetric — multimodal urban climate intelligence API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "google-genai>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "upstash-redis>=1.1.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "sse-starlette>=2.0.0",
    "supabase>=2.0.0",
    "structlog>=24.0.0",
    "firebase-admin>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
